requirements.txt
streamlit
fastapi
uvicorn
pdfplumber
chromadb
sentence-transformers
groq
python-dotenv